third_party/WebKit/LayoutTests/fast/dom/HTMLLinkElement/disabled-attribute-expected.txt

This test checks HTMLLinkElement disabled conforms to HTML5 spec for the bug id #45425 .

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS test.disabled is false
PASS test.disabled is true
PASS test.sheet.disabled is true
PASS test.sheet.disabled is false
FAIL test.disabled should be false. Was true.


PASS test_nostyle.sheet is null
PASS test_nostyle.disabled is false
FAIL test_nostyle.disabled should be false. Was true.
PASS successfullyParsed is true

TEST COMPLETE
